Move count_measurements() and changelog() to external.database.
