+++ components/server/src/database/reports.py
@@ -1,14 +1,13 @@
 """Reports collection."""
 
-from typing import Any, cast
+from typing import Any
 
 import pymongo
 from pymongo.database import Database
 
 from external.database import sessions
-from external.utils.type import Change
 from model.report import Report
-from server_utilities.functions import iso_timestamp, unique
+from server_utilities.functions import iso_timestamp
 from server_utilities.type import MetricId, ReportId, SubjectId
 
 from .filters import DOES_EXIST, DOES_NOT_EXIST
@@ -102,39 +101,3 @@ def _prepare_documents_for_insertion(
             document["delta"]["uuids"] = sorted(list(set(uuids)))
         for key, value in extra_attributes.items():
             document[key] = value
-
-
-def changelog(database: Database, nr_changes: int, **uuids):
-    """Return the changelog, narrowed to a single report, subject, metric, or source if so required.
-
-    The uuids keyword arguments may contain report_uuid="report_uuid", and one of subject_uuid="subject_uuid",
-    metric_uuid="metric_uuid", and source_uuid="source_uuid".
-    """
-    projection = {"delta": True, "timestamp": True}
-    delta_filter: dict[str, dict | list] = {"delta": DOES_EXIST}
-    changes: list[Change] = []
-    if not uuids:
-        changes.extend(
-            database.reports_overviews.find(
-                filter=delta_filter, sort=TIMESTAMP_DESCENDING, limit=nr_changes * 2, projection=projection
-            )
-        )
-    old_report_delta_filter = {f"delta.{key}": value for key, value in uuids.items() if value}
-    new_report_delta_filter = {"delta.uuids": {"$in": list(uuids.values())}}
-    delta_filter["$or"] = [old_report_delta_filter, new_report_delta_filter]
-    changes.extend(
-        database.reports.find(
-            filter=delta_filter, sort=TIMESTAMP_DESCENDING, limit=nr_changes * 2, projection=projection
-        )
-    )
-    changes = sorted(changes, reverse=True, key=lambda change: cast(str, change["timestamp"]))
-    # Weed out potential duplicates, because when a user moves items between reports both reports get the same delta
-    return list(unique(changes, get_change_key))[:nr_changes]
-
-
-def get_change_key(change: Change) -> str:
-    """Return a key for detecting equal changes."""
-    description = cast(dict[str, str], change["delta"])["description"]
-    changed_uuids = cast(dict[str, list[str]], change["delta"]).get("uuids", [])
-    key = f"{change['timestamp']}:{','.join(sorted(changed_uuids))}:{description}"
-    return key
