+++ components/server/src/external/routes/changelog.py
@@ -5,8 +5,9 @@ from typing import cast
 import bottle
 from pymongo.database import Database
 
-from database import measurements, reports
 from server_utilities.type import MetricId, ReportId, SourceId, SubjectId
+
+from ..database import measurements, reports
 
 
 def _get_changelog(database: Database, nr_changes: str, **uuids: str):
